Fix navigation menu malfunctioned
On previous code, the `isActive` would just store the boolean value for the first check of navigation menu.
However, it has been fixed by converting it into a function instead of a property.
In this way, when it get called it will checks again and again, and then returning its value.