Refactor the TOC to support on larger devices
* Hid all the TOC that are defined within header (larger devices
  will use the TOC within aside instead)
* Implemented the styles for the aside element

Note:
The TOC on larger devices will display exact on the top left of the
screen, we used fixed position and some calculation with the width
of aside element itself to make sure aside element and the main
element acts like flex box (side-by-side elements).

Hovered text still have the same effect with mobile version
when being hovered, except for the bold effect, we only
increased the weight to 500.